fix(FIT-2): Исправить ошибки // исправил ошибки
