--- src/main/java/com/project/AlphaSpringBot/AlphaSpringBotApplication.java
@@ -1,13 +1,0 @@
-package com.project.AlphaSpringBot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AlphaSpringBotApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AlphaSpringBotApplication.class, args);
-	}
-
-}
