--- src/main/java/com/project/AlphaSpringBot/botapi/BotState.java
@@ -1,18 +1,0 @@
-package com.project.AlphaSpringBot.botapi;
-
-public enum BotState {
-    START,
-    PROCESSING,
-    EAT_START,
-    EAT_NAME,
-    EAT_CALORIES,
-    EAT_FATS,
-    EAT_PROTEINS,
-    EAT_CARBOHYDRATES,
-    EAT_WEIGHT,
-    EAT_CLOSE,
-    PUKE_START,
-    PUKE_DELETE,
-    INFO,
-    PUKE_INPUT, CALORIES, CALORIES_START, CALORIES_STOP, INFO_TOTAL, EAT_YESANSWER, EAT_NOANSWER, INFO_TODAY
-}
