--- src/main/java/com/project/AlphaSpringBot/botapi/InputMessageHandler.java
@@ -1,10 +1,0 @@
-package com.project.AlphaSpringBot.botapi;
-
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
-import org.telegram.telegrambots.meta.api.objects.Message;
-
-public interface InputMessageHandler {
-    SendMessage handle(Message message);
-
-    BotState getHandlerName();
-}
