--- src/main/java/com/project/AlphaSpringBot/config/BotInit.java
@@ -1,27 +1,0 @@
-package com.project.AlphaSpringBot.config;
-
-import com.project.AlphaSpringBot.listener.TelegramBotListener;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.event.ContextRefreshedEvent;
-import org.springframework.context.event.EventListener;
-import org.springframework.stereotype.Component;
-import org.telegram.telegrambots.meta.TelegramBotsApi;
-import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
-import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
-
-@Component
-public class BotInit {
-    @Autowired
-    private TelegramBotListener bot;
-
-    @EventListener({ContextRefreshedEvent.class})
-    public void init() throws TelegramApiException {
-        TelegramBotsApi telegramBotsApi = new TelegramBotsApi(DefaultBotSession.class);
-
-        try {
-            telegramBotsApi.registerBot(bot);
-        } catch (TelegramApiException e) {
-            throw new RuntimeException(e);
-        }
-    }
-}
