--- src/main/java/com/project/AlphaSpringBot/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.project.AlphaSpringBot.repository;
-
-import com.project.AlphaSpringBot.model.User;
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserRepository extends CrudRepository<User, Long> {
-}
