--- src/main/java/com/project/AlphaSpringBot/service/ReplyMessagesService.java
@@ -1,22 +1,0 @@
-package com.project.AlphaSpringBot.service;
-
-import org.springframework.stereotype.Service;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
-
-
-@Service
-public class ReplyMessagesService {
-    private LocaleMessageService localeMessageService;
-
-    public ReplyMessagesService(LocaleMessageService messageService) {
-        this.localeMessageService = messageService;
-    }
-
-    public SendMessage getReplyMessage(long chatId, String replyMessage) {
-        return new SendMessage(String.valueOf(chatId), localeMessageService.getMessage(replyMessage));
-    }
-
-    public SendMessage getReplyMessage(long chatId, String replyMessage, Object... args) {
-        return new SendMessage(String.valueOf(chatId), localeMessageService.getMessage(replyMessage, args));
-    }
-}
