+++ src/main/java/com/project/alphaspringbot/AlphaSpringBotApplication.java
@@ -1,0 +1,13 @@
+package com.project.alphaspringbot;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AlphaSpringBotApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AlphaSpringBotApplication.class, args);
+	}
+
+}
