+++ src/main/java/com/project/alphaspringbot/botapi/BotState.java
@@ -1,0 +1,10 @@
+package com.project.alphaspringbot.botapi;
+
+public enum BotState {
+    START,
+    PROCESSING,
+    EAT_START, EAT_NAME, EAT_CALORIES, EAT_FATS, EAT_PROTEINS, EAT_CARBOHYDRATES, EAT_WEIGHT, EAT_CLOSE, EAT_YESANSWER, EAT_NOANSWER,
+    PUKE_START, PUKE_INPUT, PUKE_DELETE,
+    CALORIES, CALORIES_START, CALORIES_STOP,
+    INFO, INFO_TOTAL, INFO_TODAY
+}
