+++ src/main/java/com/project/alphaspringbot/botapi/InputMessageHandler.java
@@ -1,0 +1,10 @@
+package com.project.alphaspringbot.botapi;
+
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.api.objects.Message;
+
+public interface InputMessageHandler {
+    SendMessage handle(Message message);
+
+    BotState getHandlerName();
+}
