+++ src/main/java/com/project/alphaspringbot/config/BotConfig.java
@@ -1,0 +1,30 @@
+package com.project.alphaspringbot.config;
+
+import lombok.Data;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.MessageSource;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.support.ReloadableResourceBundleMessageSource;
+
+@Configuration
+@PropertySource("application.properties")
+@Data
+public class BotConfig {
+    @Value("${bot.name}")
+    private String botName;
+    @Value("${bot.token}")
+    private String token;
+
+    @Bean
+    public MessageSource messageSource() {
+        ReloadableResourceBundleMessageSource messageSource
+                = new ReloadableResourceBundleMessageSource();
+
+        messageSource.setBasename("classpath:messages");
+        messageSource.setDefaultEncoding("UTF-8");
+        return messageSource;
+    }
+
+}
