+++ src/main/java/com/project/alphaspringbot/config/BotInit.java
@@ -1,0 +1,27 @@
+package com.project.alphaspringbot.config;
+
+import com.project.alphaspringbot.listener.TelegramBotListener;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.event.ContextRefreshedEvent;
+import org.springframework.context.event.EventListener;
+import org.springframework.stereotype.Component;
+import org.telegram.telegrambots.meta.TelegramBotsApi;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
+
+@Component
+public class BotInit {
+    @Autowired
+    private TelegramBotListener bot;
+
+    @EventListener({ContextRefreshedEvent.class})
+    public void init() throws TelegramApiException {
+        TelegramBotsApi telegramBotsApi = new TelegramBotsApi(DefaultBotSession.class);
+
+        try {
+            telegramBotsApi.registerBot(bot);
+        } catch (TelegramApiException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
