+++ src/main/java/com/project/alphaspringbot/model/User.java
@@ -1,0 +1,103 @@
+package com.project.alphaspringbot.model;
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import java.sql.Timestamp;
+
+@Entity(name="userTestTable")
+public class User {
+
+    @Id
+    private Long chatId;
+
+    private String firstName;
+
+    private String lastName;
+
+    private String userName;
+
+    private Timestamp registredAt;
+
+    private Double growth;
+
+    private Double weight;
+
+    private Integer cCalRestriction;
+
+
+    @Override
+    public String toString() {
+        return "User{" +
+                "chatId=" + chatId +
+                ", firstName='" + firstName + '\'' +
+                ", lastName='" + lastName + '\'' +
+                ", userName='" + userName + '\'' +
+                ", registerdAt=" + registredAt +
+                '}';
+    }
+
+    public Long getChatId() {
+        return chatId;
+    }
+
+    public Double getGrowth() {
+        return growth;
+    }
+
+    public void setGrowth(Double growth) {
+        this.growth = growth;
+    }
+
+    public Double getWeight() {
+        return weight;
+    }
+
+    public void setWeight(Double weight) {
+        this.weight = weight;
+    }
+
+    public Integer getcCalRestriction() {
+        return cCalRestriction;
+    }
+
+    public void setcCalRestriction(Integer cCalRestriction) {
+        this.cCalRestriction = cCalRestriction;
+    }
+
+    public void setChatId(Long chatId) {
+        this.chatId = chatId;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getUserName() {
+        return userName;
+    }
+
+    public void setUserName(String userName) {
+        this.userName = userName;
+    }
+
+    public Timestamp getRegistredAt() {
+        return registredAt;
+    }
+
+    public void setRegistredAt(Timestamp registredAt) {
+        this.registredAt = registredAt;
+    }
+
+}
