+++ src/main/java/com/project/alphaspringbot/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package com.project.alphaspringbot.repository;
+
+import com.project.alphaspringbot.model.User;
+import org.springframework.data.repository.CrudRepository;
+
+public interface UserRepository extends CrudRepository<User, Long> {
+}
