+++ src/main/java/com/project/alphaspringbot/service/LocaleMessageService.java
@@ -1,0 +1,26 @@
+package com.project.alphaspringbot.service;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.MessageSource;
+import org.springframework.stereotype.Service;
+
+import java.util.Locale;
+
+@Service
+public class LocaleMessageService {
+    private final Locale locale;
+    private final MessageSource messageSource; //интерфейс для получения международной информации.
+
+    public LocaleMessageService(@Value("${localeTag}") String localeTag, MessageSource messageSource) {
+        this.messageSource = messageSource;
+        this.locale = Locale.forLanguageTag(localeTag);
+    }
+
+    public String getMessage(String message) {
+        return messageSource.getMessage(message, null, locale);
+    }
+
+    public String getMessage(String message, Object... args) {
+        return messageSource.getMessage(message, args, locale);
+    }
+}
