+++ src/main/java/com/project/alphaspringbot/service/ReplyMessagesService.java
@@ -1,0 +1,22 @@
+package com.project.alphaspringbot.service;
+
+import org.springframework.stereotype.Service;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+
+
+@Service
+public class ReplyMessagesService {
+    private LocaleMessageService localeMessageService;
+
+    public ReplyMessagesService(LocaleMessageService messageService) {
+        this.localeMessageService = messageService;
+    }
+
+    public SendMessage getReplyMessage(long chatId, String replyMessage) {
+        return new SendMessage(String.valueOf(chatId), localeMessageService.getMessage(replyMessage));
+    }
+
+    public SendMessage getReplyMessage(long chatId, String replyMessage, Object... args) {
+        return new SendMessage(String.valueOf(chatId), localeMessageService.getMessage(replyMessage, args));
+    }
+}
