--- src/test/java/com/project/AlphaSpringBot/AlphaSpringBotApplicationTests.java
@@ -1,13 +1,0 @@
-package com.project.AlphaSpringBot;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AlphaSpringBotApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
