+++ src/test/java/com/project/alphaspringbot/AlphaSpringBotApplicationTests.java
@@ -1,0 +1,13 @@
+package com.project.alphaspringbot;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AlphaSpringBotApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
